Add set_editable interface for Entry widget.

+++ dtk/ui/entry.py
@@ -34,6 +34,10 @@ class Entry(gtk.EventBox):
     MOVE_RIGHT = 2
     MOVE_NONE = 3
 	
+    __gsignals__ = {
+        "edit-alarm" : (gobject.SIGNAL_RUN_LAST, gobject.TYPE_NONE, ()),
+    }
+    
     def __init__(self, content="", 
                  text_color=ui_theme.get_color("entryText"),
                  text_select_color=ui_theme.get_color("entrySelectText"),
@@ -61,6 +65,7 @@ class Entry(gtk.EventBox):
         self.drag_start_index = 0
         self.drag_end_index = 0
         self.grab_focus_flag = False
+        self.editable_flag = True
         
         self.content = content
         self.cursor_index = len(self.content)
@@ -103,6 +108,17 @@ class Entry(gtk.EventBox):
         self.connect("focus-out-event", self.focus_out_entry)
         
         self.im.connect("commit", lambda im, input_text: self.commit_entry(input_text))
+        
+    def set_editable(self, editable):
+        '''Set editable.'''
+        self.editable_flag = editable
+        
+    def is_editable(self):
+        '''Whether is editable.'''
+        if not self.editable_flag:
+            self.emit("edit-alarm")
+            
+        return self.editable_flag    
         
     def set_text(self, text):
         '''Set text.'''
@@ -224,24 +240,25 @@ class Entry(gtk.EventBox):
             self.queue_draw()
             
     def backspace(self):
-        '''Backspace.'''
-        if self.select_start_index != self.select_end_index:
-            self.delete()
-        elif self.cursor_index > 0:
-            old_insert_width = self.get_content_width(self.content[0:self.cursor_index])
-            delete_char = self.get_utf8_string(self.content[0:self.cursor_index], -1)
-            self.cursor_index -= len(delete_char)
-            
-            self.content = self.content[0:self.cursor_index] + self.content[self.cursor_index + len(delete_char)::]
-            text_width = self.get_content_width(self.content)
-            insert_width = self.get_content_width(self.content[0:self.cursor_index])
-            rect = self.get_allocation()
-            if text_width < rect.width - self.padding_x * 2:
-                self.offset_x = 0
-            else:
-                self.offset_x += insert_width - old_insert_width
+        '''Backspace.'''        
+        if self.is_editable():
+            if self.select_start_index != self.select_end_index:
+                self.delete()
+            elif self.cursor_index > 0:
+                old_insert_width = self.get_content_width(self.content[0:self.cursor_index])
+                delete_char = self.get_utf8_string(self.content[0:self.cursor_index], -1)
+                self.cursor_index -= len(delete_char)
                 
-            self.queue_draw()    
+                self.content = self.content[0:self.cursor_index] + self.content[self.cursor_index + len(delete_char)::]
+                text_width = self.get_content_width(self.content)
+                insert_width = self.get_content_width(self.content[0:self.cursor_index])
+                rect = self.get_allocation()
+                if text_width < rect.width - self.padding_x * 2:
+                    self.offset_x = 0
+                else:
+                    self.offset_x += insert_width - old_insert_width
+                    
+                self.queue_draw()    
             
     def select_all(self):
         '''Select all.'''
@@ -254,7 +271,9 @@ class Entry(gtk.EventBox):
         '''Cut select text to clipboard.'''
         if self.select_start_index != self.select_end_index:
             cut_text = self.content[self.select_start_index:self.select_end_index]
-            self.delete()
+            
+            if self.is_editable():
+                self.delete()
             
             clipboard = gtk.Clipboard()
             clipboard.set_text(cut_text)
@@ -269,8 +288,9 @@ class Entry(gtk.EventBox):
     
     def paste_from_clipboard(self):
         '''Paste text from clipboard.'''
-        clipboard = gtk.Clipboard()    
-        clipboard.request_text(lambda clipboard, text, data: self.commit_entry('\\n'.join(text.split('\n'))))
+        if self.is_editable():
+            clipboard = gtk.Clipboard()    
+            clipboard.request_text(lambda clipboard, text, data: self.commit_entry('\\n'.join(text.split('\n'))))
         
     def select_to_left(self):
         '''Select to preview.'''
@@ -366,7 +386,7 @@ class Entry(gtk.EventBox):
         
     def delete(self):
         '''Delete select text.'''
-        if self.select_start_index != self.select_end_index:
+        if self.is_editable() and self.select_start_index != self.select_end_index:
             rect = self.get_allocation()
             
             self.cursor_index = self.select_start_index
@@ -619,26 +639,27 @@ class Entry(gtk.EventBox):
         
     def commit_entry(self, input_text):
         '''Entry commit.'''
-        if self.select_start_index != self.select_end_index:
-            self.delete()
+        if self.is_editable():
+            if self.select_start_index != self.select_end_index:
+                self.delete()
+                
+            self.content = self.content[0:self.cursor_index] + input_text + self.content[self.cursor_index::]
+            self.cursor_index += len(input_text)
             
-        self.content = self.content[0:self.cursor_index] + input_text + self.content[self.cursor_index::]
-        self.cursor_index += len(input_text)
-        
-        text_width = self.get_content_width(self.content)
-        rect = self.get_allocation()
-        if text_width <= rect.width - self.padding_x * 2:
-            self.offset_x = 0
-        elif self.cursor_index == len(self.content):
-            self.offset_x = text_width - (rect.width - self.padding_x * 2)
-        else:
-            old_text_width = self.get_content_width(self.content[0:self.cursor_index - len(input_text)])
-            input_text_width = self.get_content_width(input_text)
-            if old_text_width - self.offset_x + input_text_width > rect.width - self.padding_x * 2:
-                new_text_width = self.get_content_width(self.content[0:self.cursor_index])
-                self.offset_x = new_text_width - (rect.width - self.padding_x * 2)
-        
-        self.queue_draw()
+            text_width = self.get_content_width(self.content)
+            rect = self.get_allocation()
+            if text_width <= rect.width - self.padding_x * 2:
+                self.offset_x = 0
+            elif self.cursor_index == len(self.content):
+                self.offset_x = text_width - (rect.width - self.padding_x * 2)
+            else:
+                old_text_width = self.get_content_width(self.content[0:self.cursor_index - len(input_text)])
+                input_text_width = self.get_content_width(input_text)
+                if old_text_width - self.offset_x + input_text_width > rect.width - self.padding_x * 2:
+                    new_text_width = self.get_content_width(self.content[0:self.cursor_index])
+                    self.offset_x = new_text_width - (rect.width - self.padding_x * 2)
+            
+            self.queue_draw()
         
     def get_content_width(self, content):
         '''Get content width.'''
